% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.robin.R
\name{print.GxESelection}
\alias{print.GxESelection}
\title{print a GxESelection object}
\usage{
\method{print}{GxESelection}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{GxESelection object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a GxESelection object
}
\seealso{
\code{\link{GxESelection}}
}
