% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.robin.R
\name{print.roben.pred}
\alias{print.roben.pred}
\title{print a roben.pred object}
\usage{
\method{print}{roben.pred}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{roben.pred object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a roben.pred object
}
\seealso{
\code{\link{predict.roben}}
}
