% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.robin.R
\name{print.roben}
\alias{print.roben}
\title{print a roben object}
\usage{
\method{print}{roben}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{roben object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a roben object
}
\seealso{
\code{\link{roben}}
}
