% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.robin.R
\name{print.robin.pred}
\alias{print.robin.pred}
\title{print a robin.pred object}
\usage{
\method{print}{robin.pred}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{robin.pred object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a robin.pred object
}
\seealso{
\code{\link{predict.robin}}
}
