% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.robin.R
\name{print.robin}
\alias{print.robin}
\title{print a robin object}
\usage{
\method{print}{robin}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{robin object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a robin object
}
\seealso{
\code{\link{robin}}
}
