% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin-package.R
\docType{package}
\name{robin-package}
\alias{robin-package}
\title{Robust Bayesian Variable Selection for Gene-Environment Interactions}
\description{
In this package, we provide a set of robust Bayesian variable selection methods tailored for interaction analysis. A Bayesian formulation of the least absolute deviation (LAD) regression has been adopted to accommodate data contamination and long-tailed distributions in the phenotype. The default method (proposed method) conducts variable selection by accounting for structural sparsity. In particular, the spike--and--slab priors are imposed on both individual and group levels to identify important main and interaction effects (bi-level sparse-group selection).

In addition to the default method, users can also choose to use different selection structures (group-level-only or individual-level-only), methods without spike--and--slab priors or non-robust methods. In total, \emph{robin} provides 12 different methods (6 robust and 6 non-robust). Please read the "Details" below for how to configurate the method used.
}
\details{
The user friendly, integrated interface \strong{robin()} allows users to flexibly choose the fitting methods they prefer. There are three arguments in robin() that control the fitting method:
\tabular{rl}{
robust: \tab whether to use robust methods. \cr\cr
sparse: \tab whether to use the spike-and-slab priors to create sparsity. \cr\cr
structure: \tab structural identification. Three choices are available: \cr \tab "sparsegroup", "group" and “individual”.
}

robin() returns a robin object that contains the posterior estimates of each coefficients.
S3 generic functions GxESelection(), predict() and print() are implemented for robin objects.
GxESelection() takes a robin object and returns the variable selection results.
predict() takes a robin object and returns the predicted values for new observations.
}
\references{
Ren, J., Zhou, F., Li, X., Ma, S., Jiang, Y. and Wu, C. (2020). Robust Bayesian variable selection for gene-environment interactions.
}
\seealso{
\code{\link{robin}}
}
\keyword{overview}
