% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin.R
\name{robin}
\alias{robin}
\title{fit a robust Bayesian variable selection}
\usage{
robin(
  X,
  Y,
  E,
  clin = NULL,
  iterations = 10000,
  burn.in = NULL,
  robust = TRUE,
  sparse = TRUE,
  structure = c("sparsegroup", "group", "individual"),
  hyper = NULL,
  debugging = FALSE
)
}
\arguments{
\item{X}{the matrix of predictors (genetic factors) without intercept. Each row should be an observation vector. A column of 1 will be added to the X matrix
as the intercept.}

\item{Y}{the response variable. The current version of BVCfit only supports continuous response.}

\item{E}{a matrix of environmental factors for interactions.}

\item{clin}{a matrix of clinical variables. Clinical variables are not subject to penalty.}

\item{iterations}{the number of MCMC iterations.}

\item{burn.in}{the number of iterations for burn-in.}

\item{sparse}{logical flag. If TRUE, spike-and-slab priors will be used to shrink coefficients of irrelevant covariates to zero exactly. 'sparse' has effect only when VC=TRUE.}

\item{hyper}{a named list of hyperparameters.}

\item{debugging}{logical flag. If TRUE, progress will be output to the console and extra information will be returned.}

\item{structural}{logical flag. If TRUE, the coefficient functions with varying effects and constant effects will be penalized separately. 'structural' has effect only when VC=TRUE.}
}
\description{
fit a robust Bayesian variable selection model for G×E interactions.
}
\keyword{models}
